bkp/requirements.txt
pdfplumber==0.10.2
pandas==2.2.1
openpyxl==3.1.2
tabula-py==2.7.0
jpype1==1.5.0
workdays==0.4.0
reportlab==3.6.13
pyyaml==6.0.1
requests==2.31.0